[wip] Refactors File I/O
Nasty to split here over many classes. Adds an interface and starts to
move all I/O implemnetation on that interface

The interface just uses the API to sets the objects.

In this fashion, we ensure that we always have a safe in-memory
representation of objects and their relation, if the invariant are
carefully developed and tested.