Use circular mean for angle mean computation
Because the mean of (0.1,2*Pi-0.1) is zero and not Pi, we should use
the mean of the position of the angle on the unit circle, converted
back to polar coordinates.

There will be an issue if the sum of all angles gives 0 on the unit
circle. But it shouldrarely be the case as we want to estimate the
same direction.