Makes ZoneModel a Bridge
Bridge concept: Object that are set to use a pointer to an
Experiment. They provide Qt Model QObject for the Qt Widget. The slots
connections are always sets between Bridges and the Widgets. When a
new Experiment is set, it updates its internal data with the
experiment data.

This is to remove the cluttering of the ExperimentController, and
replace it by smaller simpler objects that interact together and
Signal/Slot connections.