Modifies bridge paradigm.
Previously with a change of experiment, the changes of property in
Bridge wasn't signaled, but the model were.

Now it is always reported. The only problem was with TagCloseUp
listing that would have been started/stopped several time on any
experiment change. But with a new pair of signal/slot, only notified
when a change after and edit is done, we can avoid this pitfall.