Implements a KDTree to partition the space for collision detection.
Apparently after intensive benchmarking, with random data, the
solution to use a nth_partition (quickselect) is the fastest and the
most accurates as it gives a balanced tree.

Since the tree is balanced, the variation in complexity is also
greatly reduced.

The medianEstimate is kind of not working at all

It takes omly 1.5ms to partition 10000 tags.

It takes a few hundred us to partition a more reasonable number of
ants like 100.