Fixes issues with Identifier compilation
Now just changes the unordered_map to a DenseMap.
Compilation is still usefull for concurrency issues. When concurrently
identofying, its better to save a snapshot Of identifications for that
purpose to avoid concurrent R/W access to the map.

Also its seems that the DenseMap is faster, since we will be using
values in range [0;N] for TagID. However the test are not well
designed, they should ensure that the average time is faster.