Fixes try to Load TagCloseUp for Undefined families
Closes #61

There was an Issue where fort::myrmidon::priv::TagCloseUp::Lister
could be created for fort::tags::Family::Undefined that was leading to
future issue because of thrown c++ exception.

Now the exception is raised at the creation of the Loader.

Furthermore, when adding TDD before setting up a family, was starting
the loader task in the GUI for Undefined families... Now loader task
are spin up only if a valid Tag is defined.

All new behaviors are properly unite tested