Avoids save deadlock
We use the same method for saving and saving to a new file. However
when acquiring lock for a new path, we will have an exception if we
save on the same filepathe (where we hold a lock to that file).

So we only acquire the new lock if we have no lock or we save to a
different file.