Avoids deleted reference exception with tracking solver
User may just open an experiment and compile the tracking solver and
delete the experiment. In that case we need to save in the private
TrackingSolver a reference to the identifier or we will have deleted
reference exeption when identifying things.

It may also have a problem with the zone in the solver.