Refactors Query arguments using structure
The argument are too complex not to name them properly. There are a
lot of flags, how to not mix them up ?

Also provides sensibles default for this heavy calls.

Also added a non implemented options to summarize interaction directly
from myrmidon, so the R implementation should only be a wrapper.